Split large test suite into smaller groups of tests.
